+++ library/core/src/commonMain/kotlin/com/lib/lokdroid/data/default_implementation/logger/ConsoleLogger.kt
@@ -4,5 +4,6 @@ import com.lib.lokdroid.domain.ILogger
 import com.lib.lokdroid.domain.model.Level
 
 expect object ConsoleLogger : ILogger {
+
     override fun log(level: Level, tag: String, message: String)
 }