+++ library/core/src/desktopMain/kotlin/com/lib/lokdroid/data/default_implementation/TagProvider.kt
@@ -1,5 +1,10 @@
 package com.lib.lokdroid.data.default_implementation
 
 actual object TagProvider {
-    actual fun getTag(): String = "LoKdroidDesktop"
+
+    actual fun getTag(): String {
+        val element = getTargetReferenceStackTraceElement() ?: return "????"
+
+        return element.getShortClassName()
+    }
 }